refactor: cards moved to components
